Change double to float precision and reduce nesting in PID logic

+++ src/control/pid.h
@@ -23,8 +23,8 @@
 
 struct pid
 {
-    double integral;
-    double derivative;
+    float integral;
+    float derivative;
     int32_t error;
     int32_t prev_error;
     uint32_t target;
